requirements.txt
streamlit
langchain
langchain_google_genai
langchain-community
pypdf
faiss-cpu
psycopg2